+++ lib/get.dart
@@ -1,0 +1,5 @@
+library get;
+
+export 'src/getroute.dart';
+export 'src/routes.dart';
+